requirements.txt
alembic==1.7.6
anyio==3.5.0
# Editable install with no version control (app==0.0.1)
-e /home/ivan/Documents/PythonProjects/FastAPIWeather
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.73.0
greenlet==1.1.2
h11==0.13.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.22.2
orjson==3.6.6
psycopg2-binary==2.9.3
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
timezonefinder==5.2.0
typing-extensions==4.0.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
zipp==3.7.0
